pyproject.toml
[project]
name = "call-center-ai"
description = "品驰关爱中心 AI 客服系统 - 基于阿里云 + 通义千问"
dynamic = ["version"]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  # === 核心框架 ===
  "aiohttp[speedups]~=3.10", # Async HTTP client
  "aiojobs~=1.3", # Async job scheduler
  "fastapi~=0.115", # Web framework
  "granian~=2.3", # Application server
  "jinja2~=3.1", # Template engine
  "pydantic-settings~=2.6", # Configuration management
  "pydantic[email]~=2.9", # Data validation
  "pydantic-extra-types~=2.9", # Extra types
  "python-dotenv~=1.0", # Environment variables
  "python-multipart==0.*", # Form parsing
  "pyyaml~=6.0", # YAML parser
  "structlog~=24.4", # Logging
  "tenacity~=8.2", # Retry library
  "typing-extensions~=4.12", # Typing extensions
  # === 阿里云服务 ===
  "openai~=1.55", # OpenAI SDK (DashScope 兼容 API)
  "dashscope>=1.23.4", # 阿里云 DashScope (TTS/ASR)
  # === 数据存储 ===
  "motor~=3.6", # MongoDB async driver
  "redis~=5.2", # Redis client
  # === 工具库 ===
  "phonenumbers~=8.13", # Phone number parsing
  "pytz~=2024.0", # Timezone handling
  "tiktoken~=0.8", # Tokenization
  "json-repair~=0.30", # JSON repair
  "mistune~=3.0", # Markdown parser
  "django-htmlmin~=0.11", # HTML minify
  "pyjwt~=2.9", # JWT handling
  "numpy>=2.3.5",
  "faiss-cpu>=1.13.1",
  # === LangGraph (Phase 1) ===
  "langgraph>=0.2",
]

[project.optional-dependencies]
dev = [
  "granian[reload]",           # Hot reload
  "pyright~=1.1",              # Type checker
  "pytest-asyncio~=0.24",      # Async tests
  "pytest~=8.3",               # Test framework
  "ruff~=0.7",                 # Linter
]

[tool.setuptools]
py-modules = ["app"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venv = ".venv"
venvPath = "."
